# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CRM Proto API",
    "version": "1.0.0",
    "description": "# CRM Proto API\n\nREST API для CRM системы.\n\n## Аутентификация\n\nAPI использует cookie-based аутентификацию через NextAuth.js.\nДля доступа к защищённым эндпоинтам необходимо авторизоваться через `/login`.\n\n## Ошибки\n\nВсе ошибки возвращаются в формате:\n\n```json\n{\n  \"error\": \"Описание ошибки\"\n}\n```\n\nСтандартные коды ответов:\n- `200` - Успешный запрос\n- `201` - Ресурс создан\n- `400` - Неверный запрос (ошибка валидации)\n- `401` - Не авторизован\n- `403` - Доступ запрещён\n- `404` - Ресурс не найден\n- `500` - Внутренняя ошибка сервера"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "paths": {
...
    "/api/users": {
      "post": {
        "summary": "Создать пользователя",
        "description": "Создаёт нового пользователя. Доступно только администраторам.",
        "tags": [
          "Пользователи"
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Уникальный идентификатор"
                    },
                    "email": {
                      "type": "string",
                      "description": "Email"
                    },
                    "name": {
                      "type": "string",
                      "description": "Имя"
                    },
                    "image": {
                      "type": "string"
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "admin",
                          "manager",
                          "user"
                        ],
                        "description": "Роль пользователя: admin - администратор, manager - менеджер, user - пользователь"
                      },
                      "description": "Список ролей"
                    },
                    "isActive": {
                      "type": "boolean",
                      "description": "Активен ли пользователь"
                    },
                    "lastLoginAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Дата создания"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Дата обновления"
                    }
                  },
                  "required": [
                    "id",
                    "email",
                    "name",
                    "roles",
                    "isActive",
                    "createdAt",
                    "updatedAt"
                  ],
                  "description": "Пользователь"
                },
                "example": {
                  "id": "clx1234567890abcdef",
                  "email": "manager@example.com",
                  "name": "Менеджер Продаж",
                  "role": "manager",
                  "isActive": true,
                  "createdAt": "2024-01-10T08:00:00.000Z",
                  "updatedAt": "2024-01-10T08:00:00.000Z"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "Email пользователя",
                    "format": "email"
                  },
                  "name": {
                    "type": "string",
                    "description": "Имя пользователя",
                    "minLength": 2
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6
                  },
                  "roles": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "admin",
                        "manager",
                        "user"
                      ],
                      "description": "Роль пользователя: admin - администратор, manager - менеджер, user - пользователь"
                    }
                  },
                  "googleId": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "uri"
                  }
                },
                "required": [
                  "email",
                  "name"
                ],
                "description": "Данные для создания пользователя"
              },
              "example": {
                "email": "manager@example.com",
                "name": "Менеджер Продаж",
                "password": "securePassword123",
                "role": "manager"
              }
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "summary": "Получить пользователя по ID",
        "description": "Возвращает информацию о пользователе.",
        "tags": [
          "Пользователи"
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Уникальный идентификатор"
                    },
                    "email": {
                      "type": "string",
                      "description": "Email"
                    },
                    "name": {
                      "type": "string",
                      "description": "Имя"
                    },
                    "image": {
                      "type": "string"
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "admin",
                          "manager",
                          "user"
                        ],
                        "description": "Роль пользователя: admin - администратор, manager - менеджер, user - пользователь"
                      },
                      "description": "Список ролей"
                    },
                    "isActive": {
                      "type": "boolean",
                      "description": "Активен ли пользователь"
                    },
                    "lastLoginAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Дата создания"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Дата обновления"
                    }
                  },
                  "required": [
                    "id",
                    "email",
                    "name",
                    "roles",
                    "isActive",
                    "createdAt",
                    "updatedAt"
                  ],
                  "description": "Пользователь"
                },
                "example": {
                  "id": "clx1234567890abcdef",
                  "email": "admin@example.com",
                  "name": "Администратор Системы",
                  "role": "admin",
                  "isActive": true,
                  "createdAt": "2024-01-10T08:00:00.000Z",
                  "updatedAt": "2024-01-10T08:00:00.000Z"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID пользователя"
          }
        ]
      },
      "patch": {
        "summary": "Обновить пользователя",
        "description": "Обновляет данные пользователя. Доступно только администраторам.",
        "tags": [
          "Пользователи"
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Уникальный идентификатор"
                    },
                    "email": {
                      "type": "string",
                      "description": "Email"
                    },
                    "name": {
                      "type": "string",
                      "description": "Имя"
                    },
                    "image": {
                      "type": "string"
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "admin",
                          "manager",
                          "user"
                        ],
                        "description": "Роль пользователя: admin - администратор, manager - менеджер, user - пользователь"
                      },
                      "description": "Список ролей"
                    },
                    "isActive": {
                      "type": "boolean",
                      "description": "Активен ли пользователь"
                    },
                    "lastLoginAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Дата создания"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Дата обновления"
                    }
                  },
                  "required": [
                    "id",
                    "email",
                    "name",
                    "roles",
                    "isActive",
                    "createdAt",
                    "updatedAt"
                  ],
                  "description": "Пользователь"
                },
                "example": {
                  "id": "clx1234567890abcdef",
                  "email": "admin@example.com",
                  "name": "Старший менеджер",
                  "role": "admin",
                  "isActive": true,
                  "createdAt": "2024-01-10T08:00:00.000Z",
                  "updatedAt": "2024-01-10T08:00:00.000Z"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID пользователя"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "roles": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "admin",
                        "manager",
                        "user"
                      ],
                      "description": "Роль пользователя: admin - администратор, manager - менеджер, user - пользователь"
                    }
                  },
                  "isActive": {
                    "type": "boolean"
                  },
                  "image": {
                    "type": "string",
                    "format": "uri"
                  }
                },
                "description": "Данные для обновления пользователя"
              },
              "example": {
                "name": "Старший менеджер",
                "role": "admin"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Удалить пользователя",
        "description": "Удаляет пользователя из системы. Доступно только администраторам.",
        "tags": [
          "Пользователи"
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ"
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID пользователя"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}